moves uswds components to subdirectory
